feat(backend): Dataset citations

+++ backend/src/main/kotlin/org/loculus/backend/utils/AccessionComparators.kt
@@ -7,13 +7,13 @@ typealias Version = Long
 
 object AccessionComparator : Comparator<Accession> {
     override fun compare(left: Accession, right: Accession): Int {
-        return left.compareTo(right)
+        return left.toInt().compareTo(right.toInt())
     }
 }
 
 object AccessionVersionComparator : Comparator<AccessionVersionInterface> {
     override fun compare(left: AccessionVersionInterface, right: AccessionVersionInterface): Int {
-        return when (val accessionResult = left.accession.compareTo(right.accession)) {
+        return when (val accessionResult = left.accession.toInt().compareTo(right.accession.toInt())) {
             0 -> left.version.compareTo(right.version)
             else -> accessionResult
         }
